MyGPS: Cleanup child counts on ChallengeReferral sets so they match according to later filters.

+++ src/main/java/org/studentsuccessplan/ssp/service/reference/ChallengeReferralService.java
@@ -37,7 +37,7 @@ public interface ChallengeReferralService extends
 	int getChallengeReferralCountByChallengeAndQuery(Challenge challenge,
 			String query, SortingAndPaging sAndP);
 
-	int countByChallengeIdNotOnActiveTaskList(Challenge challenge,
+	long countByChallengeIdNotOnActiveTaskList(Challenge challenge,
 			Person student, String sessionId);
 
 	List<ChallengeReferral> byChallengeIdNotOnActiveTaskList(
